requirements-dev.txt
-r requirements.txt
pytest>=8.0.0
pyinstaller>=6.0.0
pillow>=10.0.0
